requirements.txt
Django==1.9.5
django-extensions==1.6.1
https://static.dpaw.wa.gov.au/static/py/dpaw-utils/dist/dpaw-utils-0.3a11.tar.gz
Pillow==3.1.1
django-ldap-email-auth==0.1.2
South==0.8.4
dj-database-url==0.3.0
dj-static==0.0.6
djangorestframework==2.3.13
django-compressor==1.3
django-debug-toolbar==1.2
django-nose==1.2
django-leaflet==0.13.2
django-swingers==1.1.1
django-jenkins==0.15.0
django-daterange-filter==1.0.0
django-datetime-widget==0.9.2
django-flatpages-x==0.1.2
sorl-thumbnail==12.3
django-daterange-filter==1.2.0
django-datetime-widget==0.9.3
unicodecsv==0.14.1
django-storages==1.4
boto==2.39.0
arrow==0.7.0
Markdown==2.6.5
django-filter==0.12.0
openpyxl==2.3.3
